+++ h2o-algos/src/main/java/hex/tree/gbm/GBMModel.java
@@ -16,7 +16,7 @@ import water.util.SBPrintStream;
 
 import java.util.Arrays;
 
-public class GBMModel extends SharedTreeModel<GBMModel, GBMModel.GBMParameters, GBMModel.GBMOutput> 
+public class GBMModel extends SharedTreeModelGBMDRF<GBMModel, GBMModel.GBMParameters, GBMModel.GBMOutput> 
         implements Model.StagedPredictions {
 
   public static class GBMParameters extends SharedTreeModel.SharedTreeParameters {
